resource/schema: export FieldName
