+++ resource/schema/fields.go
@@ -3,7 +3,9 @@ package schema
 import (
 	"fmt"
 	"reflect"
+	"regexp"
 	"strconv"
+	"strings"
 
 	"github.com/zclconf/go-cty/cty"
 )
@@ -100,7 +102,7 @@ func Fields(target reflect.Type) FieldSet {
 			name = n
 			delete(tag, "name")
 		} else {
-			name = fieldName(f)
+			name = FieldName(f)
 		}
 		field.functag = tag["func"]
 		delete(tag, "func")
@@ -108,6 +110,23 @@ func Fields(target reflect.Type) FieldSet {
 		fields[name] = field
 	}
 	return fields
+}
+
+var reFirstCap = regexp.MustCompile("(.)([A-Z][a-z]+)")
+var reAllCap = regexp.MustCompile("([a-z0-9])([A-Z])")
+
+// FieldName returns the user-facing name of a field.
+//
+// If the field has a `name:"<fieldname>"` struct tag set, it is returned.
+// Otherwise, the field name is derived from the struct field name by
+// converting it to lower snake case.
+func FieldName(f reflect.StructField) string {
+	if n, ok := f.Tag.Lookup("name"); ok {
+		return n
+	}
+	snake := reFirstCap.ReplaceAllString(f.Name, "${1}_${2}")
+	snake = reAllCap.ReplaceAllString(snake, "${1}_${2}")
+	return strings.ToLower(snake)
 }
 
 // parseTag parses a struct tag string into a map where the key is the key of
